More stuff.

+++ install.sh
@@ -4,11 +4,13 @@ aptitude install supervisor python-pip python-gevent python-docopt python-reques
 
 DIR=$( cd "$( dirname "$0" )" && pwd )
 
-cat > $DIR/agent.conf << EOF
-[appbus]
-api_host=api-dev.appbus.io
-api_key=APIKEYHERE
-EOF
+if [ ! -f $DIR/agent.conf ]; then
+	cat > $DIR/agent.conf << EOF
+	[appbus]
+	api_host=api-dev.appbus.io
+	api_key=APIKEYHERE
+	EOF
+fi
 
 cat > /etc/supervisor/conf.d/appbus-agent.conf << EOF
 [program:appbusagent]
@@ -16,3 +18,10 @@ command=python2.7 $DIR/agent.py -f $DIR/agent.conf
 redirect_stderr=true
 stdout_logfile=/var/log/appbus.log
 EOF
+
+chmod o-rwx $DIR/agent.conf
+
+echo 
+echo 'Ready, just edit $DIR/agent.conf and set up your api key.'
+echo
+echo 'To start the agent: supervisorctl reload; supervisorctl restart appbusagent'
